denniz: sysbta: Implement all existing audio HAL versions
This is needed because libaudiohal.so can only interact with one HAL
version at a time (loading only one of the libaudiohal@version.so
implementations), so only the interface that matches the stock HAL
version will be used. libaudiohal decides which version to use by
looking through the interfaces whose variant is `default` and picking
the highest versioned one, so our hal will not interfere in the process
(because we use `sysbta` as our variant).

Signed-off-by: Nishant Kumar <www.rajsonu13@gmail.com>
Change-Id: I607d27621ffa04338d31dd05bd834548afa77d8d